added check in case book is null

+++ src/Controller/LibraryController.php
@@ -56,6 +56,11 @@ class LibraryController extends AbstractController
     {
         $book = $bookRepository->findOneBy(['isbn' => $isbn]);
         $entityManager = $manager->getManager();
+
+        if ($book === null) {
+            return $this->redirectToRoute('app_library_show_books');
+        }
+        
         $entityManager->remove($book);
         $entityManager->flush();
 
@@ -77,6 +82,11 @@ class LibraryController extends AbstractController
         $bookRepository->updateBook($data);
 
         $entityManager = $manager->getManager();
+
+        if ($book === null) {
+            return $this->redirectToRoute('app_library_show_books');
+        }
+        
         $entityManager->persist($book);
         $entityManager->flush();
